+++ Capstone.Application/Services/ProductPromotionsService/IProductPromotionsService.cs
@@ -4,7 +4,7 @@ namespace Capstone.Application.Services.ProductPromotionsService;
 
 public interface IProductPromotionsService
 {
-    Task<Result> CreateProductPromotion(string promotionId, string productId, string discountType, decimal discountValue);
+    Task<Result> CreateProductPromotion(string promotionId, string productId, string discountType, double discountValue);
 
     Task<Result<List<ProductPromotionDto>>> GetProductPromotionsByPromotionId(Guid promotionId);
