+++ Capstone.Domain/Entities/ProductPromotion.cs
@@ -6,7 +6,7 @@ public class ProductPromotion
     public Guid PromotionId { get; set; }
     public Guid ProductId { get; set; }
     public string DiscountType { get; set; } = string.Empty;
-    public decimal DiscountValue { get; set; }
+    public double DiscountValue { get; set; }
 
     public Promotion Promotion { get; set; } = null!;
     public Product Product { get; set; } = null!;
